// car-gal/interfaces.ts
export interface ICreateElementObject {
    elementTagOrElement: string | HTMLElement;
    classes?: string;
    textContent?: string;
    attr?: string[][];
    //src?: string;
    //styles?: string;
}

export interface dynCssVariables {
    id: string;
    value: string;
    childValues?: dynCssVariables[]
}

export interface IcGElementStyleObject {
    id?: string | string[];
    values?: string[][];
    childValues?: IcGElementStyleObject[];
}

export interface ICgElement {
    //get ele?: HTMLElement;
    element?: HTMLElement;
    parentElement?: HTMLElement;
    tagName?: string;
    textContent?: string;
    classes?: string;
    //styles?: string[][];
    styles?: IcGElementStyleObject;
    children?: (CgElement | ICgElement | undefined)[];
    eventListeners?: InyGalleryEventListener[];
    attr?: string[][];
    removeOnDispose?: boolean;
}

export interface InyGalleryEventListener {
    action: string;
    handler: (event: UIEvent) => void;
}
// ...
export interface IMedia {
    handler?: any;
    origin?: HTMLElement;
    element?: HTMLElement;
    containerElement: HTMLElement;
    type?: string;
    title?: string;
    description?: string;
    sizes?: ISrcSet[];
    currentSizeIndex?: number;
}

export interface ISrcSet {
    w: number;
    src: string;
}